% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embedding_utils.R
\name{check_embeddings_utils}
\alias{check_embeddings_utils}
\title{workaround for import problem on mac, get embedding utils from openAI cookbook}
\usage{
check_embeddings_utils()
}
\value{
a vector of names of functions available in this module
}
\description{
workaround for import problem on mac, get embedding utils from openAI cookbook
}
\note{
this is just an illustration of how to use basilisk with local sourced python
}
\examples{
z = check_embeddings_utils()
z
}
