% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_text.R
\name{embed_text}
\alias{embed_text}
\title{use openai embedding utility}
\usage{
embed_text(txt, model = "text-embedding-3-large")
}
\arguments{
\item{txt}{character(), will be put through paste(, collapse=" ") for submission}

\item{model}{character(1) candidate for openai.Embedding.create model parameter}
}
\value{
a list with elements ans (3072-vector of embedding coordinates) and usage (token counts)
}
\description{
use openai embedding utility
}
\examples{
# make sure key is available in env
if (interactive()) {
  txt = readLines(system.file("txt/demo.txt", package="BiocOAI"))
  tst = embed_text(txt)
  str(tst)
}
}
