% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_finefood.R
\name{extract_emb_matrix}
\alias{extract_emb_matrix}
\title{extract matrix from csv emitted by embedding process}
\usage{
extract_emb_matrix(path, col)
}
\arguments{
\item{path}{character(1) path to csv}

\item{col}{any column index selecting the embedding strings in brackets}
}
\description{
extract matrix from csv emitted by embedding process
}
\note{
uses gsub to remove brackets and coerces to numeric
}
