% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_finefood.R
\name{get_finefood_emb_matrix}
\alias{get_finefood_emb_matrix}
\title{fine food embedding as a matrix}
\usage{
get_finefood_emb_matrix()
}
\description{
fine food embedding as a matrix
}
\note{
Follows the notebook from `https://github.com/openai/openai-cookbook/blob/main/examples/Get_embeddings_from_dataset.ipynb`.  Used `text-embedding-3-small`.
}
\examples{
full = read.csv(path_finefood_emb())
mat = get_finefood_emb_matrix()
pcs = prcomp(mat)
pairs(pcs$x[,1:4], pch=19, col=full$Score, cex=.3)
}
