% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpakmeta.R
\name{labeled_vigs}
\alias{labeled_vigs}
\title{given a vector of names of packages at checked out at folder `base`,
produce a data.frame with columns `pkgname`, `views`, `vigs` and `vnc`,
which will have one row for each available Rnw or Rmd file in
the vignettes folder, and the packagename and biocViews
string in first two columns, all vignette content in
column `vigs`, and character count for each vignette in `vnc`.}
\usage{
labeled_vigs(
  base = ".",
  pks = c("DESeq2", "edgeR", "tximeta", "parody", "flowWorkspace", "Rdisop", "glmGamPoi",
    "rhdf5", "Biostrings", "affy", "snpStats")
)
}
\arguments{
\item{base}{a path to location where package sources are checked out}

\item{pks}{character() vector of package names}
}
\description{
given a vector of names of packages at checked out at folder `base`,
produce a data.frame with columns `pkgname`, `views`, `vigs` and `vnc`,
which will have one row for each available Rnw or Rmd file in
the vignettes folder, and the packagename and biocViews
string in first two columns, all vignette content in
column `vigs`, and character count for each vignette in `vnc`.
}
\note{
The biocViews are presented as colon-delimited first four
views present in DESCRIPTION.  Some packages have extravagant
collections of views.  Vignette content is restricted to 30000 characters
per vignette so that number of tokens presented is bounded.
}
