% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_finefood.R
\name{path_finefood_emb_large}
\alias{path_finefood_emb_large}
\title{path to gzipped 'large' embedding of 1000 reviews}
\usage{
path_finefood_emb_large()
}
\description{
path to gzipped 'large' embedding of 1000 reviews
}
