% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig2data.R
\name{vig2data}
\alias{vig2data}
\title{use the extract_data facility of ellmer to obtain summary information about an html
document, tailored to vignettes in bioconductor}
\usage{
vig2data(
  url = "https://bioconductor.org/packages/release/bioc/html/Voyager.html",
  maxnchar = 30000
)
}
\arguments{
\item{url}{character(1) URL for an html bioconductor vignettes}

\item{maxnchar}{numeric(1) text is truncated to a substring with this length}
}
\value{
a list with components author, topics, focused, coherence, and persuasion
}
\description{
use the extract_data facility of ellmer to obtain summary information about an html
document, tailored to vignettes in bioconductor
}
\note{
Based on code from https://cran.r-project.org/web/packages/ellmer/vignettes/structured-data.html
March 15 2025.  Requires that OPENAI_API_KEY is available in environment.
}
\examples{
if (interactive()) {
# be sure OPENAI_API_KEY is available to Sys.getenv
tst = vig2data()
str(tst)
}
}
